# Improve performance for `BaseShowTablesWithSizes` query.