## Description

`BaseShowTablesWithSizes` is executed whenever a `vttablet` process starts up.

We use `vtcombo` in our CI environment, with many different keyspaces defined in the topology or created dynamically via `CREATE DATABASE`. We noticed that when upgrading from MySQL 5.7 to MySQL 8.0 in our CI environment, the `vtcombo` startup time and the database setup times in our build jobs started to be almost twice as long as with MySQL 8.0.

I was able to trace this down to the differences in the `BaseShowTablesWithSizes` query.

The MySQL 8.0 version uses two queries combined via `UNION`. The problem here is that each of the two parts of the `UNION` ends up joining `information_schema.innodb_tablespaces`. `information_schema.innodb_tablespaces` is a MySQL system table that I believe MySQL fills on-demand with all available tablespace data, and then applies any conditions from the `WHERE` or `ON` clauses via full table scans. In short, accessing this table is expensive, and accessing it twice in separate queries doubles the cost of accessing it.

On my system, I currently have around 14k tablespaces:

```
mysql> SELECT COUNT(*) FROM information_schema.innodb_tablespaces;
+----------+
| COUNT(*) |
+----------+
|    14026 |
+----------+
1 row in set (0.18 sec)
```

Running the original query takes 0.4 seconds:

```
mysql> SELECT t.table_name,
    -> t.table_type,
    -> UNIX_TIMESTAMP(t.create_time),
    -> t.table_comment,
    -> i.file_size,
    -> i.allocated_size
    -> FROM information_schema.tables t
    -> LEFT JOIN information_schema.innodb_tablespaces i
    -> ON i.name = CONCAT(t.table_schema, '/', t.table_name) COLLATE utf8mb3_general_ci
    -> WHERE
    -> t.table_schema = database() AND not t.create_options <=> 'partitioned'
    -> UNION ALL
    -> SELECT
    -> t.table_name,
    -> t.table_type,
    -> UNIX_TIMESTAMP(t.create_time),
    -> t.table_comment,
    -> SUM(i.file_size),
    -> SUM(i.allocated_size)
    -> FROM information_schema.tables t
    -> LEFT JOIN information_schema.innodb_tablespaces i
    -> ON i.name LIKE (CONCAT(t.table_schema, '/', t.table_name, '#p#%') COLLATE utf8mb3_general_ci )
    -> WHERE
    -> t.table_schema = database() AND t.create_options <=> 'partitioned'
    -> GROUP BY
    -> t.table_schema, t.table_name, t.table_type, t.create_time, t.table_comment\G
Empty set (0.41 sec)
```

Running the query I'm proposing here takes half the time:

```
mysql> SELECT t.table_name,
    -> t.table_type,
    -> UNIX_TIMESTAMP(t.create_time),
    -> t.table_comment,
    -> SUM(i.file_size),
    -> SUM(i.allocated_size)
    -> FROM information_schema.tables t
    -> LEFT JOIN information_schema.innodb_tablespaces i
    -> ON i.name LIKE CONCAT(t.table_schema, '/', t.table_name, IF(t.create_options <=> 'partitioned', '#p#%', '')) COLLATE utf8mb3_general_ci
    -> WHERE
    -> t.table_schema = database()
    -> GROUP BY
    -> t.table_schema, t.table_name, t.table_type, t.create_time, t.table_comment
    -> \G
Empty set (0.19 sec)
```

## Related Issue(s)

* https://github.com/vitessio/vitess/issues/13374
* https://github.com/vitessio/vitess/pull/13375

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->