@arthurschreiber 2024-04-15
@shlomi-noach can you take a look? I remember you recently tried to optimize this query as well.