@arthurschreiber 2024-04-19
@shlomi-noach Thanks for taking a look! I believe this query is implicitly tested via some of the existing test cases, but I can try to find out exactly which test cases cover it.

I'll see if I can add tests that at least cover the partitioned / non partitioned table handling a bit better.