@arthurschreiber 2024-04-19
@shlomi-noach I found some existing test cases here: https://github.com/vitessio/vitess/blob/6bd23caf5596722ca4a4f64995640a86c71dc593/go/vt/vttablet/endtoend/misc_test.go#L885-L932