@arthurschreiber 2024-04-19
@shlomi-noach I pushed some light refactoring of the test case.